refactor: align package layout with public API at package root
Hoist types, exceptions, cost, reader, and core entry points to the
archivey package root; move format backends under internal/backends and
split BaseArchiveReader from the public ArchiveReader ABC. No user-facing
__all__ changes; adds package-layout-restructure OpenSpec change.
