+++ src/archivey/internal/backends/tar_reader.py
@@ -27,20 +27,20 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import BinaryIO, Iterator, Mapping, cast
 
-from archivey.internal.config import StreamConfig
-from archivey.internal.cost import (
+from archivey.cost import (
     AccessCost,
     CostReceipt,
     ListingCost,
     StreamCapability,
 )
-from archivey.internal.errors import (
+from archivey.exceptions import (
     ArchiveyError,
     CorruptionError,
     TruncatedError,
 )
+from archivey.internal.base_reader import BaseArchiveReader, ReadBackend
+from archivey.internal.config import StreamConfig
 from archivey.internal.naming import normalize_member_name
-from archivey.internal.reader import BaseArchiveReader, ReadBackend
 from archivey.internal.registry import register_reader
 from archivey.internal.streams.codecs import (
     SINGLE_FILE_CODECS,
@@ -48,7 +48,7 @@ from archivey.internal.streams.codecs import (
     open_codec_stream,
 )
 from archivey.internal.streams.streamtools import ensure_binaryio, ensure_bufferedio
-from archivey.internal.types import (
+from archivey.types import (
     ArchiveFormat,
     ArchiveInfo,
     ArchiveMember,
