+++ src/archivey/internal/naming.py
@@ -9,7 +9,7 @@ deterministic rules, while keeping the verbatim bytes in ``ArchiveMember.raw_nam
 from __future__ import annotations
 
 from archivey.internal.logs import normalization as logger
-from archivey.internal.types import MemberType
+from archivey.types import MemberType
 
 
 def normalize_member_name(decoded: str, member_type: MemberType) -> str:
