+++ src/archivey/internal/streams/lzip.py
@@ -23,7 +23,7 @@ import zlib
 from dataclasses import dataclass
 from typing import BinaryIO
 
-from archivey.internal.errors import CorruptionError, TruncatedError
+from archivey.exceptions import CorruptionError, TruncatedError
 from archivey.internal.streams.decompressor_stream import (
     SeekPoint,
     _SegmentedDecompressorStream,
