+++ tests/test_directory.py
@@ -17,7 +17,7 @@ from archivey import (
     MemberType,
     open_archive,
 )
-from archivey.internal.cost import AccessCost, ListingCost, StreamCapability
+from archivey.cost import AccessCost, ListingCost, StreamCapability
 
 # ---------------------------------------------------------------------------
 # Fixtures
@@ -438,7 +438,7 @@ def test_archive_format_named_instances() -> None:
 
 
 def test_source_name_for_path_and_stream(tmp_path: Path) -> None:
-    from archivey.internal.open_archive import source_name
+    from archivey.core import source_name
 
     p = tmp_path / "x.bin"
     p.write_bytes(b"")
