+++ tests/test_reader_contract.py
@@ -13,14 +13,14 @@ from typing import BinaryIO, Iterator
 import pytest
 
 import archivey
-from archivey.internal.cost import (
+from archivey.cost import (
     AccessCost,
     CostReceipt,
     ListingCost,
     StreamCapability,
 )
-from archivey.internal.reader import BaseArchiveReader
-from archivey.internal.types import (
+from archivey.internal.base_reader import BaseArchiveReader
+from archivey.types import (
     ArchiveFormat,
     ArchiveInfo,
     ArchiveMember,
